# api/swagger/TmsSwagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: TMSAppConnector
  description: >-
    TMSAppConnector manages ISV integrations
  contact:
    email: maheshkumar.g@varsitymgmt.com
produces:
  - application/json
basePath: /api/v1
schemes:
- "https"
...
securityDefinitions:
       Bearer:
        type: apiKey
        name: Authorization
        in: header
paths:
  /teacherHomeworkCreation:
    post:
      summary: Teacher Create Homework
      tags:
        - TeacherHomework
      operationId: createHomework
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create Homework"
        required: true
        schema:
          $ref: "#/definitions/CreationHomework"
      responses:
        200:
          description: "Successfully created Homework"
          schema:
             $ref: '#/definitions/HomeworkCreationResponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /updateHomework:
    put:
      summary: Teacher Update Homework
      tags:
        - TeacherHomework
      operationId: updateHomework
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create Homework"
        required: true
        schema:
          $ref: "#/definitions/UpdateHomework"
      responses:
        200:
          description: "Successfully created Homework"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /deleteHomework:
    put:
      summary: Teacher Delete Homework
      tags:
        - TeacherHomework
      operationId: deleteHomework
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkId"
        in: "query"
        description: "Homework Id for Deletion"
        required: true
        type: "string"
      
      responses:
        200:
          description: "Successfully Deleted Homework"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
...
  /teacherUploadingHomework:
    post:
      summary: Upload homework File by homework uuid
      operationId: uploadHomeworkByUuid
      tags:
        - TeacherHomework
      consumes:
        - multipart/form-data
      parameters:
        - in: query
          name: homeworkId
          type: integer
          required: true
        - in: query
          name: description
          type: string
          required: false
        - in: formData
          name: file[]
          type: file
          required: false
      responses:
        200:
          description: Ok
              
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      
      security:
        - Bearer: []     
...
  /getAllHomeworks:
    get:
      summary: Get All Homeworks
      tags:
        - TeacherHomework
      operationId: getAllHomeworks
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/HomeworkList'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
...
  /getStudentHomeworkEvaluation:
    get:
      summary: Student Wise Homework Evalution
      tags:
        - TeacherHomeworkEvaluation
      operationId: getStudentHomeworkEvaluation
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkId"
        in: "query"
        description: "Homework Id for Evaluation"
        required: true
        type: "integer"
        format: "int64"
      - name: "admNo"
        in: "query"
        description: "Student Admission Number"
        required: true
        type: "string"
        
        
      responses:
        200:
          description: "ok"
          schema:
             $ref: '#/definitions/StudentHomeworkEvaluation'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /getStudentHomeworkSubmissions:
    get:
      summary: Student Wise Homework Submissions 
      tags:
        - TeacherHomeworkEvaluation
      operationId: getStudentHomeworkSubmissions
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkId"
        in: "query"
        description: "Homework Id for Evaluation"
        required: true
        type: "integer"
        format: "int64"
      
      
      responses:
        200:
          description: "ok"
          schema:
             $ref: '#/definitions/StudentHomeworkSubmissions'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /studentHomeworkEvaluation:
    post:
      summary: Student Submited Homework Evaluation
      tags:
        - TeacherHomeworkEvaluation
      operationId: studentHomeworkEvaluation
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkId"
        in: "query"
        description: "Homework Id for Evaluation"
        required: true
        type: "integer"
        format: "int64"
        
      responses:
        200:
          description: "Successfully Evaluation Completed"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
...
  /setUploadedFileComments:
    post:
      summary: set Uploaded File Comments
      tags:
        - TeacherHomeworkEvaluation
      operationId: setUploadedFileComments
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkFileId"
        in: "query"
        description: "Homework File Id for Comment"
        required: true
        type: "integer"
        format: "int64"
      - name: "comments"
        in: "query"
        description: "Homework File Comment"
        required: true
        type: "string"
        
        
      responses:
        200:
          description: "Successfully Done"
          schema:
             $ref: '#/definitions/StudentUploadFileCommentResponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /deleteUploadedFileComments:
    put:
      summary: delete Uploaded File Comments
      tags:
        - TeacherHomeworkEvaluation
      operationId: deleteUploadedFileComments
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkFileCommentId"
        in: "query"
        description: "Student UploadFile Comment ID"
        required: true
        type: "integer"
        format: "int64"
        
      responses:
        200:
          description: "Successfully Done"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /setStudentHomeworkRemarks:
    post:
      summary: set Student Homework Remarks
      tags:
        - TeacherHomeworkEvaluation
      operationId: setStudentHomeworkRemarks
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "assignmentId"
        in: "query"
        description: "Student Assignment Id for Remarks"
        required: true
        type: "integer"
        format: "int64"
      - name: "remarks"
        in: "query"
        description: "Homework Remarks"
        required: true
        type: "string"
      - name: "marks"
        in: "query"
        description: "Homework marks"
        required: true
        type: number
        format: float
        
        
      responses:
        200:
          description: "Successfully Done"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /teacherPublishHomework:
    put:
      summary: Publish homework by homework uuid
      operationId: publishHomeworkByUuid
      tags:
        - TeacherHomework
      consumes:
        - multipart/form-data
      parameters:
        - in: query
          name: homeworkId
          type: integer
          required: true
      responses:
        200:
          description: Ok
              
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'  
      security:
        - Bearer: []  
...
  /createAssignmentQuestions:
    post:
      summary: creating Assignment Questions
      tags:
        - CreateOwnAssignmentTest
      operationId: createAssignmentQuestions
      parameters:
        - in: query
          name: questionId
          type: array
          items:
            type: integer
          required: true
        - in: query
          name: is_mylibrary
          type: boolean
          required: true
        - in: query
          name: assignmentSectionId 
          type: integer
          format: int64
          required: true
        
      responses:
        200:
          description: "Successfully Created Assignment Questions"
          schema:
             $ref: '#/definitions/Response'
         
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
...
  /sectionMarks:
    put:
      summary: section Marks
      tags:
        - CreateOwnSection
      operationId: sectionMarks
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: sectionId
          type: integer
          format: int64
          required: true
        - in: query
          name: positiveMarks
          type: number
          format: float
          required: true
        - in: query
          name: negativeMarks
          type: number
          format: float
          required: true  
          
        
      responses:
        200:
          description: Ok
          
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error' 
      security:
        - Bearer: []                     
  /deleteSection:
    delete:
      summary: delete section for adding questions 
      tags:
        - CreateOwnSection
      operationId: delete Section
      parameters:
        - in: query
          name: Id
          type: integer
          required: true          
      responses:
        200:
          description: success
          
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error' 
      security:
        - Bearer: []
  /getSections:
    get:
      summary: Get sections
      tags:
        - CreateOwnSection
      operationId: assignmentSections
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: homeWorkId
          type: integer
          format: int64
          required: true 
        
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/SectionList'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'       
      security:
        - Bearer: []
  /questionMarks:
    put:
      summary: Questionmarks
      tags:
        - CreateOwnAssignmentTest
      operationId: marksQuestion
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "sectionId"
        in: "query"
        required: true
        type: "integer"
        format: "int64" 
      - name: "positiveMarks"
        in: "query"
        required: true
        type: "string"
      - name: "negativeMarks"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionMarksresponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /questionDelete:
    put:
      summary: questionDelete
      tags:
        - CreateOwnAssignmentTest
      operationId: questionDelete
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "questionId"
        in: "query"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionDeleteresponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /getAssignmentfileComments:
    get:
      summary: Assignment file Comments
      tags:
        - TeacherHomeworkEvaluation
      operationId: assignment file comments
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: query
          name: assignmentfileId
          type: integer
          required: true
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/fileComments'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /deletefilecomments:
    delete:
      summary: delete file wise comments
      tags:
        - TeacherHomeworkEvaluation
      operationId: delete comments
      parameters:
        - in: query
          name: assignmentfileId
          type: integer
          required: true        
      responses:
        200:
          description: success
          
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error' 
      security:
        - Bearer: []
  /getStudentResponseAssignmentTest:
    get:
      summary: take studnet assignment test
      operationId: getStudentResponseAssignmentTest
      tags:
        - AssignmentTestResponse
      parameters:
        - in: query
          name: studentAssignmentId
          type: "integer"
          format: "int64"
          required: true
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: "#/definitions/assignmentTestPaper"
      security:
        - Bearer: []    
  /getHomeworkAssignmentTest:
    get:
      summary: take homework assignment test
      operationId: getHomeworkAssignmentTest
      tags:
        - AssignmentTestResponse
      parameters:
        - in: query
          name: homeworkId
          type: "integer"
          format: "int64"
          required: true
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: "#/definitions/assignmentTestPaper"
      security:
        - Bearer: []
...
  /deleteHomeworkFile:
    put:
      summary: Teacher Delete Homework File
      tags:
        - TeacherHomework
      operationId: deleteHomeworkFile
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "homeworkfileId"
        in: "query"
        description: "Homework file Id for Deletion"
        required: true
        type: "string"
      
      responses:
        200:
          description: "Successfully Deleted Homework"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []  
definitions:
...
  QuestionMarksresponse:
    type: object
    properties:
      message:
        type: string
        default: " marks added to question Sucessfully"
      status:
        type: string
        default: "Sucess"
...
  Response:
    type: "object"
    properties:
      success:
          type: string
      message:
          type: string
  HomeworkCreationResponse:
    type: "object"
    properties:
      success:
          type: string
      message:
          type: string
      homeworkId:
          type: integer
          format: int64
  StudentUploadFileCommentResponse:
    type: "object"
    properties:
      success:
          type: string
      message:
          type: string
      uploadFileCommentId:
          type: integer
          format: int64
...
  Error:
    description: Schema for API errors.
    type: object
    properties:
      status:
        type: integer
        description: original HTTP error code, should be consistent with the response
          HTTP code
        minimum: 100
        maximum: 599
      code:
        type: string
        description: 'The error code, in screaming snake case: eg PAYLOAD_VALIDATION_ERROR'
        pattern: "^[A-Z]+(_[A-Z]+)*$"
      message:
        type: string
        description: descriptive error message for debugging
      moreInfo:
        type: string
        format: uri
        description: link to documentation to investigate further and finding support
      details:
        type: array
        description: list of problems causing this error
        items:
          title: errorDetail
          description: schema for specific error cause
          type: object
          properties:
            field:
              type: string
              description: JSON object notation specifying the element in the request
                payload causing the error. eg company.entitlements[3]
            code:
              type: string
              description: 'The error code detail in screaming snake case: eg MISSING_VALUE'
              pattern: "^[A-Z]+(_[A-Z]+)*$"
            message:
              type: string
              description: descriptive error detail message for debugging
            moreInfo:
              type: string
              format: uri
              description: link to documentation to investigate further and finding support
                for error detail
          required:
            - code
    required:
      - status
      - code